prior_work/total_requirements.txt
absl-py==2.1.0
accelerate==1.5.2
affine==2.4.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
appnope==0.1.4
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
attrs==25.1.0
audioread==3.0.1
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.13.3
bitsandbytes==0.42.0
blinker==1.9.0
Brotli==1.1.0
build==1.2.2.post1
buildozer==1.5.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
codecarbon==2.8.3
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
Cython==3.0.12
dataclasses-json==0.6.7
datasets==3.2.0
debugpy==1.8.12
decorator==5.1.1
Deprecated==1.2.18
dill==0.3.8
distlib==0.3.9
distro==1.9.0
docutils==0.21.2
durationpy==0.9
et_xmlfile==2.0.0
executing==2.2.0
faiss-cpu==1.10.0
Faker==36.1.1
fastapi==0.115.8
fief-client==0.20.0
filelock==3.17.0
filetype==1.2.0
fiona==1.10.1
flatbuffers==25.1.24
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.9.0
gast==0.4.0
geopandas==1.0.1
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.70.0
h11==0.14.0
h5py==3.12.1
hf-xet==1.0.3
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.27.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
inflate64==1.0.1
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.31.0
jax==0.4.30
jaxlib==0.4.30
jedi==0.19.2
Jinja2==3.1.5
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jwcrypto==1.5.6
kafka-python==2.0.6
keras==2.12.0
Kivy==2.3.1
Kivy-Garden==0.1.5
kivy-garden.xcamera==2020.613
kivymd==1.2.0
kiwisolver==1.4.8
kubernetes==32.0.0
langchain==0.3.21
langchain-chroma==0.2.1
langchain-community==0.3.20
langchain-core==0.3.45
langchain-huggingface==0.1.2
langchain-mistralai==0.2.6
langchain-text-splitters==0.3.7
langsmith==0.3.4
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.44.0
lxml==5.3.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.0
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.3.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
multivolumefile==0.2.3
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.24.2
nest-asyncio==1.6.0
networkx==3.4.2
npm==0.1.1
NREL-PySAM==6.0.1
numba==0.61.0
numpy==1.26.4
nvidia-ml-py==12.570.86
oauthlib==3.2.2
onnxruntime==1.20.1
openai==1.66.3
opencv-python==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
opt_einsum==3.4.0
optional-django==0.1.0
optree==0.14.0
orjson==3.10.15
overrides==7.7.0
OWSLib==0.34.dev0
packaging==24.2
pandas==2.2.3
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
plotly-express==0.4.1
pluggy==1.5.0
pooch==1.8.2
posthog==3.11.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==5.29.4
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py7zr==0.22.0
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybcj==1.0.3
pycparser==2.22
pycryptodomex==3.21.0
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
pyet==1.3.1
Pygments==2.19.1
pyngrok==7.2.4
pynvml==12.0.0
pyogrio==0.10.0
pyparsing==3.2.1
pypdf==5.3.1
PyPDF2==3.0.1
PyPika==0.48.9
pyppmd==1.1.1
pyproj==3.7.1
pyproject_hooks==1.2.0
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-for-android==2024.1.21
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.0
pyzstd==0.16.2
questionary==2.1.0
RapidFuzz==3.11.0
rasterio==1.4.3
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.1
sentence-transformers==3.4.1
sh==1.14.3
shapely==2.0.7
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.6
soxr==0.5.0.post1
SQLAlchemy==2.0.37
stack-data==0.6.3
starlette==0.45.3
statsmodels==0.14.4
streamlit==1.43.0
sympy==1.13.3
tenacity==9.0.0
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.12.0
tensorflow-metal==1.0.0
termcolor==2.3.0
texttable==1.7.0
tf_keras==2.16.0
thefuzz==0.22.1
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.21.0
toml==0.10.2
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
typer==0.15.1
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.30.0
watchfiles==1.0.4
wcwidth==0.2.13
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
wrapt==1.14.1
xarray==2025.1.2
xxhash==3.5.0
yarl==1.18.3
yaspin==3.1.0
zipp==3.21.0
zstandard==0.23.0
